requirements.txt
appdirs==1.4.4
asgiref==3.3.1
Authlib==0.15.2
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
dataclasses==0.6
Django==3.1.4
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.2
openapi-codec==1.3.2
packaging==20.8
pathspec==0.8.1
pycparser==2.20
pyparsing==2.4.7
pytz==2020.4
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
